Change channel/role position and bitrate type to `u32` (#2149)

+++ src/model/guild/audit_log/change.rs
@@ -93,8 +93,8 @@ pub enum Change {
     },
     /// Voice channel bitrate was changed.
     Bitrate {
-        old: Option<u64>,
-        new: Option<u64>,
+        old: Option<u32>,
+        new: Option<u32>,
     },
     /// Channel for invite code or guild scheduled event was changed.
     ChannelId {
@@ -294,8 +294,8 @@ pub enum Change {
     },
     /// Channel or role position was changed.
     Position {
-        old: Option<u64>,
-        new: Option<u64>,
+        old: Option<u32>,
+        new: Option<u32>,
     },
     /// Preferred locale of a guild was changed.
     PreferredLocale {
@@ -720,7 +720,7 @@ impl<'de> Visitor<'de> for ChangeVisitor {
             Available: bool,
             AvatarHash: String,
             BannerHash: String,
-            Bitrate: u64,
+            Bitrate: u32,
             ChannelId: ChannelId,
             Code: String,
             Colour: u64,
@@ -761,7 +761,7 @@ impl<'de> Visitor<'de> for ChangeVisitor {
             OwnerId: UserId,
             PermissionOverwrites: Vec<PermissionOverwrite>,
             Permissions: Permissions,
-            Position: u64,
+            Position: u32,
             PreferredLocale: String,
             PrivacyLevel: u64,
             PruneDeleteDays: u64,
